feat: Реализовал слой DAO и разделил Storage
