--- src/main/java/com/yourcompany/gym/dao/Storage.java
@@ -1,44 +1,0 @@
-package com.yourcompany.gym.dao;
-
-import com.yourcompany.gym.model.Trainee;
-import com.yourcompany.gym.model.Trainer;
-import com.yourcompany.gym.model.Training;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import jakarta.annotation.PostConstruct;
-import java.util.HashMap;
-import java.util.Map;
-
-@Component
-public class Storage<V> {
-    private final Map<Long, V> storage = new HashMap<>();
-
-    private final Map<Long, Trainee> trainees = new HashMap<>();
-    private final Map<Long, Trainer> trainers = new HashMap<>();
-    private final Map<Long, Training> trainings = new HashMap<>();
-
-
-    @Value("${initial.data.path}")
-    private String initialDataPath;
-
-
-    @PostConstruct
-    public void initialize() {
-
-        System.out.println("LOG: Initializing storage from file specified by path: " + initialDataPath);
-
-    }
-
-    public Map<Long, Trainee> getTrainees() {
-        return trainees;
-    }
-
-    public Map<Long, Trainer> getTrainers() {
-        return trainers;
-    }
-
-    public Map<Long, Training> getTrainings() {
-        return trainings;
-    }
-}
